+++ lib/hobbit.rb
@@ -1,3 +1,4 @@
 require 'hobbit/base'
+require 'hobbit/request'
 require 'hobbit/response'
 require 'hobbit/version'